Adding pagination, authentication & authorization

+++ backend/routes/posts.js
@@ -3,6 +3,7 @@ const multer = require('multer');
 
 const Post = require('../models/post');
 const routes = express.Router();
+const checkAuth = require('../middleware/check-auth');
 
 const MIME_TYPE_MAP = {
 	'image/png': 'png',
@@ -26,12 +27,13 @@ const storage = multer.diskStorage({
 	}
 });
 
-routes.post('', multer({ storage: storage }).single('image'), (req, res, next) => {
+routes.post('', checkAuth, multer({ storage: storage }).single('image'), (req, res, next) => {
 	const url = req.protocol + '://' + req.get('host');
 	const post = new Post({
 		title: req.body.title,
 		content: req.body.content,
-		imagePath: url + '/images/' + req.file.filename
+		imagePath: url + '/images/' + req.file.filename,
+		creator: req.userData.userId
 	});
 	post.save().then((result) => {
 		res.status(201).json({
@@ -46,7 +48,7 @@ routes.post('', multer({ storage: storage }).single('image'), (req, res, next) =
 	});
 });
 
-routes.put('/:id', multer({ storage: storage }).single('image'), (req, res, next) => {
+routes.put('/:id', checkAuth, multer({ storage: storage }).single('image'), (req, res, next) => {
 	let imagePath = req.body.imagePath;
 	if (req.file) {
 		const url = req.protocol + '://' + req.get('host');
@@ -56,24 +58,42 @@ routes.put('/:id', multer({ storage: storage }).single('image'), (req, res, next
 		_id: req.params.id,
 		title: req.body.title,
 		content: req.body.content,
-		imagePath: imagePath
+		imagePath: imagePath,
+		creator: req.userData.userId
 	});
-	console.log(post);
-	Post.updateOne({ _id: req.params.id }, post).then((result) => {
-		console.log(result);
-		res.status(200).json({
-			message: 'Update Successful!'
-		});
+	Post.updateOne({ _id: req.params.id, creator: req.userData.userId }, post).then((result) => {
+		if (result.nModified > 0) {
+			res.status(200).json({
+				message: 'Update Successful!'
+			});
+		} else {
+			res.status(401).json({
+				message: 'Not Authorized!'
+			});
+		}
 	});
 });
 
 routes.get('', (req, res, next) => {
-	Post.find().then((documents) => {
-		res.status(200).json({
-			message: 'Posts fetched successfully',
-			posts: documents
+	const pageSize = +req.query.pageSize;
+	const currentPage = +req.query.page;
+	const postQuery = Post.find();
+	let fetchedPosts;
+	if (pageSize && currentPage) {
+		postQuery.skip(pageSize * (currentPage - 1)).limit(pageSize);
+	}
+	postQuery
+		.then((documents) => {
+			fetchedPosts = documents;
+			return Post.count();
+		})
+		.then((count) => {
+			res.status(200).json({
+				message: 'Posts fetched successfully',
+				posts: fetchedPosts,
+				maxPosts: count
+			});
 		});
-	});
 });
 
 routes.get('/:id', (req, res, next) => {
@@ -88,12 +108,17 @@ routes.get('/:id', (req, res, next) => {
 	});
 });
 
-routes.delete('/:id', (req, res, next) => {
-	Post.deleteOne({ _id: req.params.id }).then((result) => {
-		console.log(result);
-		res.status(200).json({
-			message: 'Post Deleted'
-		});
+routes.delete('/:id', checkAuth, (req, res, next) => {
+	Post.deleteOne({ _id: req.params.id, creator: req.userData.userId }).then((result) => {
+		if (result.n > 0) {
+			res.status(200).json({
+				message: 'Post Deleted'
+			});
+		} else {
+			res.status(401).json({
+				message: 'Not Authorized!'
+			});
+		}
 	});
 });
